+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Splash Screen NPC start ammo
 - Decrease minimum fire distance for Gun weapons for NPC
 - Fix EMPTY / NO MAGS spam #44 (add reload / fire button down event) 
+- Additional particles when bullet hits player shields
+- Fix bullet bouncing on hit with damage #36 
 
 ## 1.0.6